+++ packages/hightide-design/src/semantic-token-resolvers/with-appearance.ts
@@ -1,4 +1,5 @@
 import type { HexColorToken } from '../primitive-tokens/color'
+import type { ColorPairToken } from '../theme-tokens/theme-tokens-config'
 import type { ThemeTokens } from '../theme-tokens/theme-tokens'
 import { HexColorUtils } from '../utils/hex'
 
@@ -6,13 +7,13 @@ export type Appearance = 'normal' | 'subtle' | 'faded'
 
 export const resolveWithAppearance = (params: {
   themeTokens: ThemeTokens,
-  color: HexColorToken,
+  colorPair: ColorPairToken,
   appearance: Appearance,
 }): HexColorToken => (
   HexColorUtils.blend(
-    params.themeTokens.color.surface.color,
+    params.colorPair.color,
     HexColorUtils.hexWithAlpha(
-      params.color,
+      params.colorPair.onColor,
       params.themeTokens.config.appearancePercentages[params.appearance]
     )
   )
@@ -20,22 +21,22 @@ export const resolveWithAppearance = (params: {
 
 export const resolveAsFaded = (params: {
   themeTokens: ThemeTokens,
-  color: HexColorToken,
+  colorPair: ColorPairToken,
 }): HexColorToken => (
   resolveWithAppearance({
     themeTokens: params.themeTokens,
-    color: params.color,
+    colorPair: params.colorPair,
     appearance: 'faded',
   })
 )
 
 export const resolveAsDescription = (params: {
   themeTokens: ThemeTokens,
-  color: HexColorToken,
+  colorPair: ColorPairToken,
 }): HexColorToken => (
   resolveWithAppearance({
     themeTokens: params.themeTokens,
-    color: params.color,
+    colorPair: params.colorPair,
     appearance: 'subtle',
   })
 )